fix warnings from debug compile

+++ cpp/include/cudf/io/parquet.hpp
@@ -790,7 +790,7 @@ class parquet_writer_options {
   void set_max_page_size_bytes(size_t size_bytes)
   {
     CUDF_EXPECTS(size_bytes >= 1024, "The maximum page size cannot be smaller than 1KB.");
-    CUDF_EXPECTS(size_bytes <= std::numeric_limits<int32_t>::max(),
+    CUDF_EXPECTS(size_bytes <= static_cast<size_t>(std::numeric_limits<int32_t>::max()),
                  "The maximum page size cannot exceed 2GB.");
     _max_page_size_bytes = size_bytes;
   }
@@ -831,7 +831,7 @@ class parquet_writer_options {
    */
   void set_max_dictionary_size(size_t size_bytes)
   {
-    CUDF_EXPECTS(size_bytes <= std::numeric_limits<int32_t>::max(),
+    CUDF_EXPECTS(size_bytes <= static_cast<size_t>(std::numeric_limits<int32_t>::max()),
                  "The maximum dictionary size cannot exceed 2GB.");
     _max_dictionary_size = size_bytes;
   }
@@ -1379,7 +1379,7 @@ class chunked_parquet_writer_options {
   void set_max_page_size_bytes(size_t size_bytes)
   {
     CUDF_EXPECTS(size_bytes >= 1024, "The maximum page size cannot be smaller than 1KB.");
-    CUDF_EXPECTS(size_bytes <= std::numeric_limits<int32_t>::max(),
+    CUDF_EXPECTS(size_bytes <= static_cast<size_t>(std::numeric_limits<int32_t>::max()),
                  "The maximum page size cannot exceed 2GB.");
     _max_page_size_bytes = size_bytes;
   }
@@ -1420,7 +1420,7 @@ class chunked_parquet_writer_options {
    */
   void set_max_dictionary_size(size_t size_bytes)
   {
-    CUDF_EXPECTS(size_bytes <= std::numeric_limits<int32_t>::max(),
+    CUDF_EXPECTS(size_bytes <= static_cast<size_t>(std::numeric_limits<int32_t>::max()),
                  "The maximum dictionary size cannot exceed 2GB.");
     _max_dictionary_size = size_bytes;
   }
